requirements.txt
streamlit
langchain
openai
pinecone-client
langchain-openai
langchain-pinecone
langchain-community
python-dotenv
requests
tiktoken
beautifulsoup4
lxml
